update
added double

+++ src/main/ValueComparator.java
@@ -3,10 +3,10 @@ package main;
 import java.util.Comparator;
 import javafx.util.Pair;
 
-public class ValueComparator implements Comparator<Pair<Integer,Integer>> {
+public class ValueComparator implements Comparator<Pair<Integer,Double>> {
 
   @Override
-  public int compare(Pair<Integer,Integer> o1, Pair<Integer,Integer> o2) {
+  public int compare(Pair<Integer,Double> o1, Pair<Integer,Double> o2) {
     return o1.getValue().compareTo(o2.getValue());
   }
 }
